Re-design and implement Volition C API (#40)
# Volition C API (Take Two)

The API design in PR #39 was relatively high-level with lots of
simplifications made to the Protobuf message hierarchy.

This PR instead implements a slightly lower-level API with the primary
goal being making the API more consistent and easier to mechanically
translate without sacrificing type safety. The new design follows the
Protobuf native C++ API more closely.

## API Design

The API design makes extensive use of opaque handle types. The
implementations are not exposed to the consumer of the API. Type safety
and class invariants are preserved as the user can only interact with
the API objects through pre-defined methods.

### API Objects

The primary API objects include:

| Type | Description |
| :- | :- |
| `unary_optimizer_t` | Represents a connection endpoint to the
optimizer service. |
| `start_optimization_command_t` | Contains the problem definition and
various parameters to be passed to the optimizer. |
| `optimizer_generated_query_stream_t` | The query stream returned by
the optimizer service. |
| `optimizer_generated_query_t` | A single query generated by the
optimizer. |
| `simulation_evaluation_completed_response_t` | Client-generated
evaluation result corresponding to an evaluation request from the
optimizer. |

## Misc Changes
- Enable various [hardening
options](https://best.openssf.org/Compiler-Hardening-Guides/Compiler-Options-Hardening-Guide-for-C-and-C++.html)
and sanitizer flags for GCC and Clang builds.
- Add `.clang-format` and `.clang-tidy` files.

## Next Steps
- Move the API implementation code to another directory.
- Update the reference client example to use pre-built Volition SDK
artifacts.